backend/requirements.txt
# Document Processing
beautifulsoup4==4.12.2
pypdf2==3.0.1
python-docx==1.1.0
lxml==5.1.0

# Text Extraction
pdfplumber==0.10.3

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Data Processing
python-dotenv==1.0.0

# AWS SageMaker
boto3==1.34.0
